requirements.txt
cached-property==1.5.1
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
docker==3.5.1
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
freeze==1.0.10
idna==2.6
jsonschema==2.6.0
Markdown==3.0.1
Pillow==5.3.0
pytz==2018.6
PyYAML==3.13
requests==2.18.4
six==1.11.0
texttable==0.9.1
urllib3==1.22
websocket-client==0.53.0
